requirements.txt
discord.py
openai
schedule